requirements.txt
# Core
pydantic>=2.0
python-dotenv>=1.0

# Data loading and parsing
pandas>=2.0

# Database
psycopg[binary]>=3.1
pgvector>=0.2.0

# LangGraph and LLM (for merge/same-or-new agent)
langgraph>=0.2
langchain-core>=0.3
langchain-openai>=0.2
# langchain-anthropic>=0.2  # optional

# Embeddings (pick one or both)
openai>=1.0
# sentence-transformers>=2.0  # optional, for local embeddings

# Vector store (if not using pgvector for identity embeddings)
# lancedb>=0.5
# faiss-cpu>=1.0

# Geocoding (optional: set GEOCODE_ENABLED=1 to fill lat/lon from addresses)
geopy>=2.4

# CLI / scripting
typer>=0.12

# API for frontend (optional: POST /ingest with CSV)
fastapi>=0.109
uvicorn[standard]>=0.27
